[REFACTOR] : 챗뷰델과 메시지뷰모델로 분리
