+++ app/src/main/java/com/example/frompet/ui/chat/ChatViewModel.kt
@@ -1,6 +1,5 @@
 package com.example.frompet.ui.chat
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -14,87 +13,31 @@ import com.google.firebase.database.ValueEventListener
 import com.google.firebase.firestore.FirebaseFirestore
 
 class ChatViewModel : ViewModel() {
-    private val _chatMessages = MutableLiveData<List<ChatMessage>>()
-    val chatMessages: LiveData<List<ChatMessage>> get() = _chatMessages
-    private val _isTyping = MutableLiveData<Boolean>()
-    val isTyping: LiveData<Boolean> get() = _isTyping
-    private val _lastMessages = HashMap<String, MutableLiveData<ChatMessage?>>()
-    private val _newMessages = MutableLiveData<HashMap<String, Boolean>>()
-    val newMessages: LiveData<HashMap<String, Boolean>> get() = _newMessages
+
+    private val _lastChats = HashMap<String, MutableLiveData<ChatMessage?>>()
+    private val _newChats = MutableLiveData<HashMap<String, Boolean>>()
+    val newChats: LiveData<HashMap<String, Boolean>> get() = _newChats
 
 
     private val database = FirebaseDatabase.getInstance().reference
-    private val firestore = FirebaseFirestore.getInstance()
     private val auth = FirebaseAuth.getInstance()
 
     fun chatRoom(uid1: String, uid2: String): String {
         return if (uid1 > uid2) "$uid1+$uid2" else "$uid2+$uid1" //두 사람 채팅에는 항상 합친 동일한 구분자로 생성함
     }
 
-    fun getLastMessageLiveData(chatRoomId: String): LiveData<ChatMessage?> {
-        return _lastMessages.getOrPut(chatRoomId) { MutableLiveData<ChatMessage?>() }
+    fun lastChatLiveData(chatRoomId: String): LiveData<ChatMessage?> {
+        return _lastChats.getOrPut(chatRoomId) { MutableLiveData<ChatMessage?>() }
     }
 
-    fun sendMessage(receiverId: String, message: String) {
-        val currentUserId = auth.currentUser?.uid ?: return
-        val chatRoomId = chatRoom(currentUserId, receiverId)
-
-        firestore.collection("User").document(currentUserId).get()
-            .addOnSuccessListener { document ->
-                val currentUser = document.toObject(User::class.java)
-                val senderPetName = currentUser?.petName ?: "Unknown"
-
-                val chatMessage = ChatMessage(
-                    senderId = currentUserId,
-                    senderPetName = senderPetName,
-                    receiverId = receiverId,
-                    message = message,
-                    timestamp = System.currentTimeMillis()
-                )
-                database.child("chatMessages").child(chatRoomId).push().setValue(chatMessage)
-                database.child("lastMessages").child(chatRoomId).setValue(chatMessage)
-                database.child("newMessages").child(chatRoomId).child(receiverId).setValue(true)
-                    .addOnSuccessListener {
-                        loadPreviousMessages(chatRoomId)
-                    }.addOnFailureListener { exception ->
-                        Log.d("jun", "메시지전송실패: ${exception.message}")
-                    }
-            }.addOnFailureListener { exception ->
-                Log.d("jun", "연결실패: ${exception.message}")
-            }
-    }
-    fun sendImage(chatMessage: ChatMessage) {
-        val chatRoomId = chatRoom(chatMessage.senderId, chatMessage.receiverId)
-
-        database.child("chatMessages").child(chatRoomId).push().setValue(chatMessage)
-        database.child("lastMessages").child(chatRoomId).setValue(chatMessage)
-        database.child("newMessages").child(chatRoomId).child(chatMessage.receiverId).setValue(true)
-    }
-
-
-    fun loadPreviousMessages(chatRoomId: String) {
-        database.child("chatMessages").child(chatRoomId)
-            .addValueEventListener(object : ValueEventListener {
-                override fun onDataChange(snapshot: DataSnapshot) {
-                    val messages =
-                        snapshot.children.mapNotNull { it.getValue(ChatMessage::class.java) }
-                    _chatMessages.postValue(messages)
-                }
-
-                override fun onCancelled(databaseError: DatabaseError) {
-                    Log.d("jun", "불러오기실패: ${databaseError.message}")
-                }
-            })
-    }
-
-    fun loadLastMessage(currentUserId: String, otherUserId: String) { //  리얼타임 베이스의 구조를 최적화하여 필요한 데이터만 읽으려고 라스트메시지 노드 따로 추가함
+    fun loadLastChats(currentUserId: String, otherUserId: String) { //  리얼타임 베이스의 구조를 최적화하여 필요한 데이터만 읽으려고 라스트메시지 노드 따로 추가함
         val chatRoomId = chatRoom(currentUserId, otherUserId)
 
         database.child("lastMessages").child(chatRoomId)
             .addValueEventListener(object : ValueEventListener {
                 override fun onDataChange(snapshot: DataSnapshot) {
                     val message = snapshot.getValue(ChatMessage::class.java)
-                    _lastMessages[chatRoomId]?.value = message
+                    _lastChats[chatRoomId]?.value = message
                 }
 
                 override fun onCancelled(databaseError: DatabaseError) {}
@@ -106,7 +49,7 @@ class ChatViewModel : ViewModel() {
         database.child("newMessages").child(chatRoomId).child(currentUserId).setValue(false)
     }
 
-    fun loadNewMessages() {
+    fun loadNewChats() {
         val currentUserId = auth.currentUser?.uid ?: return
         database.child("newMessages")
             .orderByChild(currentUserId)
@@ -118,7 +61,7 @@ class ChatViewModel : ViewModel() {
                             key to (it.child(currentUserId).getValue(Boolean::class.java) ?: false)
                         }
                     }.toMap()
-                    _newMessages.value = HashMap(newMessageRooms)
+                    _newChats.value = HashMap(newMessageRooms)
                 }
 
                 override fun onCancelled(databaseError: DatabaseError) {}
@@ -149,21 +92,4 @@ class ChatViewModel : ViewModel() {
         })
     }
 
-
-    fun checkTypingStatus(receiverId: String) {
-        database.child("typingStatus").child(receiverId)
-            .addValueEventListener(object : ValueEventListener {
-                override fun onDataChange(snapshot: DataSnapshot) {
-                    val status = snapshot.getValue(Boolean::class.java) ?: false
-                    _isTyping.value = status
-                }
-
-                override fun onCancelled(databaseError: DatabaseError) {}
-            })
-    }
-
-    fun setTypingStatus(isTyping: Boolean) {
-        val currentId = auth.currentUser?.uid ?: return
-        database.child("typingStatus").child(currentId).setValue(isTyping)
-    }
 }
